+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 - added French translation (thanks to Wronschien)
 - removed hard-coded fish and piranha setup and moved to Lua instead
 - removed the limit of at most 8 fish/piranha shoals per level
+- removed hard-coded small Cobra radius setup and moved to Lua instead
 - fixed a crash with old custom levels that have sectors pointing to invalid floor data (#5568)